+++ src/common_test.rs
@@ -1,4 +1,4 @@
-use std::f64::consts::FRAC_1_SQRT_2;
+use std::f32::consts::FRAC_1_SQRT_2;
 
 use nalgebra::{Complex, DVector, dvector};
 
@@ -32,7 +32,7 @@ pub fn double_sub<
         .h(2)
         .h(3);
 
-    const L: usize = 5; //(std::f64::consts::PI * 2f64.sqrt() / 4f64).floor() as usize;
+    const L: usize = 5; //(std::f32::consts::PI * 2f32.sqrt() / 4f32).floor() as usize;
 
     for _ in 0..L {
         circuit = circuit.call("sub", 0);
@@ -128,7 +128,7 @@ pub fn deep_sub<D: BuildSimulator<HybridCircuit> + DebuggableSimulator + StoredC
         forward_steps += 1;
     }
 
-    let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+    let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
     correct[0] = cart!(FRAC_1_SQRT_2);
     correct[1 << (LEVELS - 1)] = cart!(FRAC_1_SQRT_2);
 
@@ -141,7 +141,7 @@ pub fn deep_sub<D: BuildSimulator<HybridCircuit> + DebuggableSimulator + StoredC
         }
         assert_eq!(forward_steps, backward_steps);
 
-        let mut correct = DVector::<Complex<f64>>::zeros(1 << LEVELS);
+        let mut correct = DVector::<Complex<f32>>::zeros(1 << LEVELS);
         correct[0] = cart!(1);
         assert!(equal_to_matrix_c(sim.current_state(), &correct, 0.001));
     }
